+++ Resources/doc/CHANGELOG.md
@@ -3,6 +3,7 @@ Netgen Open Graph Bundle changelog
 
 ## 1.0.1 (22.04.2015)
 
+* Remove need for specifying image alias in `ezimage` field type handler
 * Fix issue with being able to add only one handler of the same type
 
 ## 1.0 (21.11.2014)